FEATURE: moved code to service class
